+++ CHANGELOG.md
@@ -6,6 +6,7 @@ GAMSPy 0.13.1
 - General
   - Support output redirection for NEOS backend.
   - Support GAMSPy to GAMS automatic conversion.
+  - Add support for old way of installing a license. 
 - Documentation
   - Update model documentation to show how to redirect NEOS output to a file.
 - Testing